Models/requirements.txt


plotly
python-dateutil
pandas==2.2.3
more_itertools==10.7.0
python-dotenv==1.1.0
Requests==2.32.3
elasticsearch==8.12.1
fastapi==0.115.12
numpy<2
pydantic==2.11.4
pytest==8.3.5
joblib==1.5.0
mlflow==2.22.0
scikit_learn==1.6.1
xgboost==3.0.1
prometheus-client
mysql-connector-python
email-validator>=2.0
